fix postgres duration queries to return expected integer type

+++ src/db/postgres/sql/topic-get-content-by-id.sql
@@ -1,7 +1,7 @@
 -- topic including content
 SELECT *,
-	extract(epoch FROM lease_seconds_preferred) AS lease_seconds_preferred,
-	extract(epoch FROM lease_seconds_min) AS lease_seconds_min,
-	extract(epoch FROM lease_seconds_max) AS lease_seconds_max
+	extract(epoch FROM lease_seconds_preferred)::integer AS lease_seconds_preferred,
+	extract(epoch FROM lease_seconds_min)::integer AS lease_seconds_min,
+	extract(epoch FROM lease_seconds_max)::integer AS lease_seconds_max
 FROM topic
 WHERE id = $(topicId)
